pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "flask-cors>=5.0.1",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openai>=1.72.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.3",
    "uvicorn>=0.34.0",
    "requests>=2.32.3",
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.2",
    "werkzeug>=3.1.3",
    "trafilatura>=2.0.0",
    "numpy>=2.2.4",
]
